+++ game/game.hpp
@@ -163,8 +163,8 @@ public:
 
     void set_lives(uint8_t lives);
     void set_score(uint16_t score);
-    void set_tile(int x, int y, Tile tile);
-    Tile get_tile(int x, int y);
+    void set_tile(uint8_t x, uint8_t y, Tile tile);
+    Tile get_tile(uint8_t x, uint8_t y);
 
     int get_grid_width() const {
         return grid_width;
